requirements.txt
streamlit
pillow
numpy
opencv-python
matplotlib
